feat: allow artifact selection on `RolePage`
The grid format is still around, but you can now select an artifact on a page! I just look at the path given after the role-page's path. I use that as the id in the map where I later grab the artifact from. The other role-pages are a bit skiwampus at the moment, but that will be fixed in time. I discovered that I have to use `[[...artifactPaths]]` so that it will display the original role-page as well as the role-page with path parameters from the same javascript file. Hopefully some of this makes sense: it's late.