feat: :art: allow sticky footer
Using a technique from Ben Frain's book, I set the body to fill 100% and then had the main content grow to fill all the space except for the footer. That way, the footer sticks to the bottom, but can go even farther if the content needs it to. Note that I had to style the `#__next` element to let this technique work. It shouldn't cause problems, but I thought I'd mention it.