feat: changed artifact-cards-area to use flexbox instead of grid.
This allows the last item to be more balanced instead of off to one side. The artifact cards are a fixed width, but they can wrap onto as many lines as they need. I needed to get rid of the margin in order to get the alignment to work correctly. I had to restrain the this-website thumbnail because it is wider than it is tall. OH, I renamed `artifact-cards` to `artifact-cards-area` to avoid more confusion.